Initial Changes for the EPaper Windows Application
Added in the base Arduino Driver for controlling the EPaper module, this contains basic serial connection via USB using a hard coded connection to the COM8 port (my Arduino Mega lives there). The communication is basic Serial commands with the structure <CommandCode:int><CommandArguments>

Added in a minor shared project to contains the overlap needed by both the applications and the Arduino Driver's managed part. This right now just contains a duplication of the error codes that live in the unmanaged part of the driver

Added in a very basic WPF UI application that gives the user access to a drop down that allows the user to be able to send commands to the Arduino through the driver and C++\CLI.

Added minor changes to my library for my RGB desk toy
* Added the base arduino project
* Added the classes that control a single RGB LED
* Added the implementation for the glow and other light change functions